Create SignIn and SignUp componentes and add to each page and remove hunter pages
