+++ src/pages/dev/signin.tsx
@@ -1,48 +1,21 @@
 import {
-  Link as ChakraLink,
   Flex,
   Heading,
   Icon,
   IconButton,
   Image,
   SimpleGrid,
-  Text,
   VStack
 } from '@chakra-ui/react'
-import { FiEye, FiEyeOff } from 'react-icons/fi'
 
-import { Button } from 'components/Button'
 import { Container } from 'components/Container'
 import { FaLongArrowAltLeft } from 'react-icons/fa'
 import Head from 'next/head'
-import Link from 'next/link'
 import Router from 'next/router'
-import { TextInput } from 'components/TextInput'
-import { signInSchema } from 'utils/yupSchemas'
-import { useForm } from 'react-hook-form'
-import { useState } from 'react'
-import { yupResolver } from '@hookform/resolvers/yup'
+import { SignInForm } from 'components/Form/SignInForm'
 
-interface SignInData {
-  email: string
-  password: string
-}
-
-export default function SignIn() {
-  const [showPassword, setShowPassword] = useState(false)
+export default function SignInPage() {
   const highlightColor = 'blue.500'
-
-  const {
-    register,
-    handleSubmit,
-    formState: { isSubmitting, errors }
-  } = useForm({
-    resolver: yupResolver(signInSchema)
-  })
-
-  const onSubmit = async (values: SignInData) => {
-    console.log(values)
-  }
 
   return (
     <>
@@ -68,14 +41,7 @@ export default function SignIn() {
         w="100vw"
         h="calc(100vh - 5rem)"
       >
-        <SimpleGrid
-          as="form"
-          onSubmit={handleSubmit(onSubmit)}
-          columns={2}
-          columnGap={40}
-          maxW="992px"
-          mx="auto"
-        >
+        <SimpleGrid columns={2} columnGap={40} maxW="992px" mx="auto">
           <VStack align="center" justify="center">
             <Image
               src="/img/logo.png"
@@ -87,48 +53,7 @@ export default function SignIn() {
               Encontre e ajude seus amigos devs
             </Heading>
           </VStack>
-          <VStack align="flex-start" maxW="20rem" spacing="5">
-            <Heading as="h3" color={highlightColor}>
-              Login para Devs
-            </Heading>
-            <VStack w="100%" spacing="1">
-              <TextInput
-                inputName="email"
-                error={errors.email}
-                type="email"
-                placeholder="Email"
-                {...register('email')}
-              />
-              <TextInput
-                inputName="password"
-                error={errors.password}
-                rightIcon={
-                  showPassword ? (
-                    <Icon
-                      as={FiEyeOff}
-                      onClick={() => setShowPassword(false)}
-                    />
-                  ) : (
-                    <Icon as={FiEye} onClick={() => setShowPassword(true)} />
-                  )
-                }
-                type={showPassword ? 'text' : 'password'}
-                placeholder="Password"
-                {...register('password')}
-              />
-            </VStack>
-            <Button type="submit" fullWidth={true} isLoading={isSubmitting}>
-              Entrar
-            </Button>
-            <Text as="span" fontSize="sm" color="gray.500">
-              Não tem conta?
-            </Text>
-            <Link href="/dev/signup" passHref={true}>
-              <ChakraLink color={highlightColor} fontSize="sm">
-                Cadastre-se
-              </ChakraLink>
-            </Link>
-          </VStack>
+          <SignInForm />
         </SimpleGrid>
       </Container>
     </>
